move side-functions around
